requirements.txt
openai>=1.0.0
requests>=2.25.0
Pillow>=8.0.0
matplotlib>=3.3.0
pandas>=1.3.0
python-dotenv>=0.19.0